return cache values for non-arrays where required

+++ lib/class/channel.class.php
@@ -613,17 +613,17 @@ class Channel extends database_object implements media, library_item
                         while ($this->strtohex(substr($clchunk, 0, 4)) == "4F676753") {
                             $hex                = $this->strtohex(substr($clchunk, 0, 27));
                             $ogg_nr_of_segments = hexdec(substr($hex, 26 * 2, 2));
-                            if ((substr($clchunk, 27 + $ogg_nr_of_segments + 1, 6) == "vorbis") || (substr($clchunk, 27 + $ogg_nr_of_segments, 4) == "Opus")) {
-                                $hex .= $this->strtohex(substr($clchunk, 27, $ogg_nr_of_segments));
+                            if ((substr($clchunk, (int) (27 + $ogg_nr_of_segments + 1), 6) == "vorbis") || (substr($clchunk, (int) (27 + $ogg_nr_of_segments), 4) == "Opus")) {
+                                $hex .= $this->strtohex(substr($clchunk, 27, (int) $ogg_nr_of_segments));
                                 $ogg_sum_segm_laces = 0;
                                 for ($segm = 0; $segm < $ogg_nr_of_segments; $segm++) {
                                     $ogg_sum_segm_laces += hexdec(substr($hex, 27 * 2 + $segm * 2, 2));
                                 }
-                                $this->header_chunk .= substr($clchunk, 0, 27 + $ogg_nr_of_segments + $ogg_sum_segm_laces);
+                                $this->header_chunk .= substr($clchunk, 0, (int) (27 + $ogg_nr_of_segments + $ogg_sum_segm_laces));
                                 if (strlen((string) $clchunk) < (27 + $ogg_nr_of_segments + $ogg_sum_segm_laces)) {
                                     $this->header_chunk_remainder = (int) (27 + $ogg_nr_of_segments + $ogg_sum_segm_laces - strlen((string) $clchunk));
                                 }
-                                $clchunk = substr($clchunk, 27 + $ogg_nr_of_segments + $ogg_sum_segm_laces);
+                                $clchunk = substr($clchunk, (int) (27 + $ogg_nr_of_segments + $ogg_sum_segm_laces));
                             } else {
                                 // no more interesting headers
                                 $clchunk = '';
